Pulling classes to public files and start clearing. Migrating code-generating to Attr class new behaviour. More to come
