+++ .gitignore
@@ -1,7 +1,4 @@
-
-
-/Users/engmoht/IdeaProjects/HtmlTags/src/main/resources/ignored/comments
-
+src/main/resources/ignored
 ##############################
 ## Java
 ##############################
